+++ CHANGELOG.md
@@ -16,8 +16,10 @@ The following sections are considered for each release: **Added, Changed, Fixed,
 - Change width of columns on the backlog screen, so there is no overflow on smaller screens
 - Now shows invalid external identifiers in the atomic data entry form, these are highlighted in red.
 
-### Added
+### Removed
+- Remove nondefining properties from product edit terms screen
 
+### Added
 - Handling of non-defining properties with coded values no longer in the FHIR CodeSystem -
   highlights these non-defining properties in red to attract user attention
 
@@ -34,6 +36,7 @@ The following sections are considered for each release: **Added, Changed, Fixed,
 ### Changed
 
 - Artg Id search to only appear for AMT branches
+- Add product level nondefining properties to brand and pack
 - Dose form type field for NMPC from optional to mandatory as requested.
 - Disable delete option for robot created product saved data